Asinhrona komunikacija (rabbitmq) na transportu

+++ Transport/src/main/java/com/nwtProject/model/Transport.java
@@ -15,7 +15,7 @@ public class Transport {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
     @Column (name = "idTransporta")
-    private int idTransporta;
+    private long idTransporta;
 
     @Column(name = "cijenaMax")
     private Double cijenaMax;
@@ -56,11 +56,11 @@ public class Transport {
                 idTransporta, cijenaMax, datum, idAutoprevoznika, status, idKompanije, naziv);
     }
 
-	public int getIdTransporta() {
+	public long getIdTransporta() {
 		return idTransporta;
 	}
 
-	public void setIdTransporta(int idTransporta) {
+	public void setIdTransporta(long idTransporta) {
 		this.idTransporta = idTransporta;
 	}
